@@ -13,6 +13,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Throwable;
 
 /**
  * @Route("/admin/feed")
@@ -22,10 +23,10 @@ use Symfony\Component\Routing\Annotation\Route;
 class FeedCrudController extends AbstractController
 {
     /** @var FeedProcessor */
-    private $feedProcessor;
+    private FeedProcessor $feedProcessor;
 
     /** @var Permissions */
-    private $permissions;
+    private Permissions $permissions;
 
     public function __construct(FeedProcessor $feedProcessor, Permissions $permissions)
     {
@@ -62,7 +63,7 @@ class FeedCrudController extends AbstractController
             // Fetch feed, but fail gracefully
             try {
                 $this->feedProcessor->fetchFeeds([$feed], true);
-            } catch (\Throwable $exception) {
+            } catch (Throwable $exception) {
                 // Do nothing
             }
 
