@@ -6,14 +6,14 @@ namespace App\DataFixtures;
 use App\Entity\Feed;
 use App\Repository\RssUserRepository;
 use Doctrine\Bundle\FixturesBundle\Fixture;
-use Doctrine\Common\Persistence\ObjectManager;
+use Doctrine\Persistence\ObjectManager;
 
 class AppFixtures extends Fixture
 {
     /**
      * @var RssUserRepository
      */
-    private $userRepository;
+    private RssUserRepository $userRepository;
 
     public function __construct(RssUserRepository $userRepository)
     {
