@@ -25,7 +25,7 @@ class Feed
      *
      * @ORM\Column(type="text")
      */
-    private $feedUrl;
+    private string $feedUrl;
 
     /**
      * @var FeedItem[]|ArrayCollection
@@ -40,32 +40,32 @@ class Feed
      *
      * @ORM\Column(type="text", nullable=true)
      */
-    private $title;
+    private ?string $title;
 
     /**
      * @var string|null
      *
      * @ORM\Column(type="text", nullable=true)
      */
-    private $description;
+    private ?string $description;
 
     /**
      * @var DateTimeInterface|null
      *
      * @ORM\Column(type="text", nullable=true)
      */
-    private $icon;
+    private ?string $icon;
 
     /**
      * @ORM\Column(type="datetimetz", nullable=true)
      */
-    private $lastModified;
+    private ?DateTimeInterface $lastModified;
 
     /**
      * @ORM\ManyToOne(targetEntity="App\Entity\RssUser", inversedBy="feeds")
      * @ORM\JoinColumn(nullable=false)
      */
-    private $rssUser;
+    private RssUser $rssUser;
 
     public function __construct()
     {
