@@ -9,6 +9,7 @@ use App\Repository\FeedItemRepository;
 use App\Repository\FeedRepository;
 use DateInterval;
 use DateTime;
+use Exception;
 use FaviconFinder\Favicon;
 use FeedIo\Feed as FeedIoFeed;
 use FeedIo\Feed\ItemInterface as RawFeedItem;
@@ -32,28 +33,28 @@ class Processor
     private const FEED_ITEM_BATCH_SIZE = 5;
 
     /** @var FeedIo */
-    private $feedIo;
+    private FeedIo $feedIo;
 
     /** @var FeedRepository */
-    private $feedRepository;
+    private FeedRepository $feedRepository;
 
     /** @var ClientInterface */
-    private $guzzle;
+    private ClientInterface $guzzle;
 
     /** @var FeedItemRepository */
-    private $feedItemRepository;
+    private FeedItemRepository $feedItemRepository;
 
     /** @var LoggerInterface */
-    private $logger;
+    private LoggerInterface $logger;
 
     /** @var string */
-    private $readabilityEndpoint;
+    private string $readabilityEndpoint;
 
     /** @var Favicon */
-    private $faviconFinder;
+    private Favicon $faviconFinder;
 
     /** @var int */
-    private $ingestedCount = 0;
+    private int $ingestedCount = 0;
 
     public function __construct(
         FeedIo $feedIo,
@@ -79,7 +80,7 @@ class Processor
      * @param array|Feed[] $feeds
      * @param bool         $bypassLastModified
      *
-     * @throws \Exception
+     * @throws Exception
      */
     public function fetchFeeds(array $feeds, bool $bypassLastModified): void
     {
